requirements-gpu.txt
https://download.pytorch.org/whl/cu117/torch-1.13.0%2Bcu117-cp310-cp310-linux_x86_64.whl
https://download.pytorch.org/whl/cu117/torchvision-0.14.0%2Bcu117-cp310-cp310-linux_x86_64.whl
numpy==1.22.3
tqdm==4.62.3
scikit-learn==1.1.3
onnxruntime-gpu==1.13.1
pandas==1.4.3
pysam==0.20.0
opencv-python==4.6.0.66
Pillow==9.3.0
seaborn==0.12.2
matplotlib==3.6.2
pre-commit==2.21.0
